+++ hitbox/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [0.1.1] 2021-05-30
+### Changed
+- Downgrade actix dependency to 0.10
 
 ## [0.1.0] - 2021-05-29
 ### Added